handle empty stream

+++ src/main/scala/example/Throughput.scala
@@ -20,7 +20,7 @@ object Throughput {
       ssc, kafkaParams, topics)
 
     stream.foreachRDD { rdd =>
-      println(rdd.map(_._2.size).fold(0L)(_+_))
+      println(rdd.map(_._2.size.toLong).fold(0L)(_+_))
     }
 
     ssc.start()
